Fixing various signedness warnings

+++ tools/solidityUpgrade/UpgradeChange.cpp
@@ -27,7 +27,10 @@ using namespace solidity::tools;
 
 void UpgradeChange::apply()
 {
-	m_source.replace(m_location.start, m_location.end - m_location.start, m_patch);
+	m_source.replace(
+		static_cast<size_t>(m_location.start),
+		static_cast<size_t>(m_location.end - m_location.start), m_patch
+	);
 }
 
 void UpgradeChange::log(bool const _shorten) const
@@ -56,7 +59,7 @@ void UpgradeChange::log(bool const _shorten) const
 	string line;
 	while (getline(output, line))
 	{
-		os << string(leftpad, ' ');
+		os << string(static_cast<size_t>(leftpad), ' ');
 		AnsiColorized(os, true, {formatting::BOLD, formatting::BLUE}) << "| ";
 		AnsiColorized(os, true, {}) << line << endl;
 	}
